backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
contourpy==1.2.0
cryptography==41.0.5
cycler==0.12.1
exceptiongroup==1.1.3
fastapi==0.104.1
fonttools==4.44.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
kiwisolver==1.4.5
mssql==1.0.1
mysql-connector-python==8.2.0
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyodbc==5.0.1
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
